pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.5.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [ { include = "gendiff" }, ]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
wemake-python-styleguide = "^0.16.1"
pytest = "^7.1.3"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
